Update governing-the-digital-layer.md
The current draft is not legally accurate. I have attempted to work within the structure to improve legal accuracy. In particular: (1) I left the references to the Charter but tried to place it in context. The Charter does not govern private sector organizations. It does not expressly protect privacy even in respect of governmental institutions. (2) There is a vast difference between public sector and private sector privacy laws. Consent is not a requirement under public sector laws. (3) The omission of any reference to MFIPPA is glaring since the City of Toronto and municipal corps (like Toronto Hydro) are governed by MFIPPA. (4) A key challenge (and, therefore, opportunity) for digital stewardship structures is that they provide a medium through which the various tensions and differences between public sector, private sector and health sector laws can be reconciled. This is not brought out in the current draft.