Update business-model-options.md
I would be pleased to discuss this page with you. It is problematic because it does not quite capture how these structures would fit into the Canadian legal framework. I have tried to improve it but it is still open to criticism. (1) The law governing the business model can be different from that governing the end user. E.g. a not-for-profit may NOT (probably will NOT) be subject to PIPEDA. However, contributors and end users will be subject to FIPPA, MFIPPA, PIPEDA, PHIPA as the case may be. There will be recourse to the OPC and IPC. The "trust" just complicates things.  (2) The for-profit model does not avoid scrutiny any more than the not-for-profit will mean scrutiny. Arguably, the for-profit is under GREATER scrutiny because it will be clearly subject to PIPEDA. I have previously posted a blog on this point. (3) There is no reason why governments, citizens, local businesses, private sector in general can't access data in the private sector model. Think of the Google Cloud Public Dataset Program. .